pyproject.toml
[project]
name = "history-agent-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.1",
    "django-cors-headers>=4.9.0",
    "django-ninja>=1.5.2",
    "django-simple-captcha>=0.6.3",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "openai>=2.14.0",
    "pillow>=12.1.0",
    "psycopg[binary]>=3.3.2",
    "pydantic[email]>=2.12.5",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "uvicorn>=0.40.0",
]
